inventory/requirements.txt
fastapi==0.75.0
redis-oms=0.0.20